Fix columns width when wrapping text
Currently when table's width is above terminal width, all column widths will be recalculated and the final result is equal column widths. This leads to unwanted behavior for our use-case. for example:
General | Usage
Android | some very long command with options
In this case after wrapping the text, both columns will have the same width, which will lead to output like this:
General                          |                               Usage
Android                          |         some very long command with options

###Proposed solution
Use each column's required width and calculate its relative size based on current terminal width. For example if we have the following structure:
Column A | Column B | Column C
   30                  50                20
and the available terminal width is 80, the calculation of the columns should be:
Column A = (30 / 100) * 80 = 24
Column B = (50 / 100) * 80 = 40
Column C = (20 / 100) * 80  = 16
Column A + Column B + Column C = 80

###Extended solution
In most of our use-cases, we have 2 columns in our table - the first one contains short text (around 10 symbols) and the second column has some very long text. After applying suggested solution, first column's width is reduced and the text inside it is trimmed or placed on two rows. Add minColumnWidth property and if the width of the column is below this value, do not reduce columns size. In case the calculated value is below minColumnWidth, return minColumnWidth. The problem here is in the following scenario:
Column A | Column B | Column C | Column D | Column E
    10                   20                30                 10                 30
The sum is 100. Max terminal width is 80.
Lets set the minColumnWidth property to 23. Here's the calculation:
Column A - 10 < 25 => 10
Column B - (20 / 100) * 80 = 16
Column C - (30 / 100) * 80 = 24
Column D - 10 < 25 => 10
Column E - (30 / 100) * 80 = 24
A + B + C +D +E = 74
But some of the columns are less then already defined minColumnWidth(23). If we set them to minColumnWidth the values are:
Column A - 10 < 23 => 10
Column B - (20 / 100) * 80 = 16 < 23 => 23
Column C - (30 / 100) * 80 = 24
Column D - 10 < 23 => 10
Column E - (30 / 100) * 80 = 24
A + B + C +D +E = 91

This way we've exceeded terminal width. The proposed solution is to modify the formula based on the remaining terminal and the remaining required size.
For each column we check if its size is below minColumnWidth. In case it is, we set this column as static (its size will not change) and we remove the width
from remaining terminal and remaining required width. If the width is above minColumnWidth we execute:
(required_col_width / remaining_required_size) * remaining_terminal_width
In case the value is above minColumnWidth, we save it as dynamic size (could be changed later).
In case the value is below minColumnWidth, we save it as static size and remove it from remaining_terminal_width. The original value should be removed from remaining_required_size.
At this point we have to recalculate all dynamic values and apply the same logic as above.