Add recalculateColWidths method
In cases when we recalculate column widths, we must be sure that new column width is at least min_column_width(for all columns, which are above min_column_width, after calculation we do not want to have value less than min_column_width - that's the point of this parameter).
Add warning in the case when calculated table width is above terminal width (it is possible when min_column_width has high value).