Split string correctly
Remove usage of `js-beautify` module for spliting the string for multiline usage. There are two problems with `js-beautify` module:
* text that contains <...> is always split on several lines. `js-beautify` considers <...> as HTML tag and places is to new line.
* `js-beautify` does not guarantee that the string will be split based on wrap_line_length value. The string is split based on the first delimiter AFTER wrap_line_length.

### Proposed solution
Use own function instead of `js-beautify` to split the lines. The following rules are applied:
* if content is below wrap_line_length, return the content without modifying it
* if content is above wrap_line_length the rules are:
    * if the symbol on index "wrap_line_length" is delimiter, take all symbols up to it and split the rest of the text based on this rules (call the method with rest of the text)
    * if the next symbol is not delimiter, try to find the last delimiter in the symbols up to wrap_line_length index.
        * in case delimiter is found, split the line by it and split everything after it based on the same rules
        * in case delimiter is not found, split the word, this means - take all symbols up to wrap_line_length and split the rest of the string based on the same rules.

### 'Code' problem
If the passed text is colorized, the string contains several specific colorization symbols. They are rendered by the console as colors, but when checking string.length, they are calculated (typically the length of these symbols is 8-9). After we split the content on several lines, we lose the colorization of subsequent lines.

### 'Code' problem solution
When splitting the text, calculate the length of all "code" symbols which we can include until we reach the allowed length of the column. When splitting, check the number of "code" segments in the line - if they are odd number, get the last "code" segment and add it to the beginning of the remaining text. This way we have the correct colorization on the next lines.