Fix "code" segment shown on new line
In cases when we want to split the text inside column of cli-table, we try to find out the first delimiter. As this delimiter may be part of "code" segment, we must be sure to include the whole segment on the current line.

Also fix the case when the text starts with delimiter and should be splitted, for example '-configuration'. The code finds the last delimiter, which is on index 0 and for the next iteration it uses `remaining = content.substring(index).tirm()`, which leads to infinite loop.