Fix issue with padding
In case there's padding on each cell, in some cases cli-table trims the text in the cell (replacing the parts that could not be shown with ....). This is due to incorrect value passed to extendWrapLine method, where the padding of the cell is not taken into account. Make sure correct padding is used.